pdf-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
Pillow==10.2.0
python-dotenv==1.0.0